requirements.txt
Django==3.1.3
psycopg2-binary>=2.8
psycopg2==2.8.6
Pillow==8.1.2
djangorestframework==3.12.2
pyyaml==5.4.1
uritemplate==3.0.1
drf-yasg==1.20.0
python-decouple==3.4
dj-database-url==0.5.0
django-sslserver==0.22
django-dbbackup
django-cors-headers
djoser
djangorestframework_simplejwt==4.6.0
social-auth-app-django==4.0.0